requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorchvideo>=0.1.5
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0
yt-dlp>=2023.7.0
av>=10.0.0
